requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
Flask-Mail==0.9.1
python-dotenv==1.0.0
APScheduler==3.10.4
openpyxl==3.1.2
reportlab==4.0.7
WTForms==3.1.1
email-validator==2.1.0
gunicorn==21.2.0
